auth_service/requirements.txt
Flask==3.0.0
PyJWT==2.8.0
python-json-logger==2.0.7
gunicorn==21.2.0
